Добавляет решение задания Модуляция
