+++ js/data.js
@@ -1,0 +1,56 @@
+const OBJECTS_COUNT = {
+  MIN: 0,
+  MAX: 26
+};
+
+const COMMENTS_COUNT = {
+  MIN: 0,
+  MAX: 30
+};
+
+const AVATARS_COUNT = {
+  MIN: 0,
+  MAX: 6
+};
+
+const SENTENCES_COUNT = {
+  MIN: 0,
+  MAX: 4
+};
+
+const LIKES_COUNT = {
+  MIN: 15,
+  MAX: 200
+};
+
+const RANDOM_VALUE = {
+  MIN: 0,
+  MAX: Number.MAX_VALUE
+};
+
+const PHOTO_DESCRIPTIONS = ['Закат над океаном',
+  'Густой лес в тумане', 'Старая каменная башня', 'Поле с подсолнухами', 'Заснеженные горы на рассвете',
+  'Уличное кафе в Париже', 'Тихий пляж с белым песком', 'Пейзаж с речкой и мостиком', 'Яркий фейерверк ночью',
+  'Городская улица в дождь', 'Цветущий сад весной', 'Корабль в открытом море', 'Тропа в осеннем лесу',
+  'Кошка на подоконнике', 'Ночное небо с луной', 'Живописный водопад', 'Рынок со свежими фруктами',
+  'Древние руины в пустыне', 'Скала у побережья', 'Группа воздушных шаров в небе', 'Кофе и книги на столе',
+  'Пастбище с коровами', 'Детский парк с аттракционами', 'Чайный домик в Японии', 'Маяк на скалистом берегу',
+  'Озеро с отражением леса', 'Снежная деревня зимой', 'Песчаные дюны в пустыне', 'Романтический мост в вечернем свете',
+  'Улица с цветными домами', 'Яхты в маленькой гавани', 'Кристально чистое горное озеро', 'Поле с маковыми цветами',
+  'Маленькая хижина в горах', 'Закат в тропиках', 'Стадо оленей на рассвете', 'Городской парк с фонтаном',
+  'Облачный горизонт над мегаполисом', 'Гребной канал с лодками', 'Старинный замок в горах'];
+
+const MESSAGES = ['Всё отлично! В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.', 'Как можно было поймать такой неудачный момент?!'];
+
+
+const NAMES = ['Алексей', 'Мария', 'Иван', 'Екатерина', 'Дмитрий', 'Ольга', 'Николай', 'Анна', 'Сергей', 'Елена',
+  'Владимир', 'Татьяна', 'Артур', 'Наталья', 'Павел', 'Ирина', 'Кирилл', 'Светлана', 'Максим', 'Юлия'];
+
+export {
+  OBJECTS_COUNT, COMMENTS_COUNT, AVATARS_COUNT, SENTENCES_COUNT, LIKES_COUNT, RANDOM_VALUE,
+  PHOTO_DESCRIPTIONS, MESSAGES, NAMES
+};
